refactor(Ladder): 원시값 래핑 객체 Height를 제거하고 Ladder로 책임을 위임
