+++ src/main/java/ladder/domain/Ladder.java
@@ -8,14 +8,40 @@ import ladder.domain.randomGenerator.RungGenerator;
 
 public class Ladder {
 
+    private static final int MIN_HEIGHT_RANGE = 1;
+    private static final int MAX_HEIGHT_RANGE = 100;
+
     private final List<Floor> floors;
 
-    public Ladder(Height height, int personCount, RungGenerator rungGenerator) {
+    public Ladder(String inputHeight, int personCount, RungGenerator rungGenerator) {
+        int height = parseAndValidateHeight(inputHeight);
         floors = makeFloors(height, personCount, rungGenerator);
     }
 
-    private List<Floor> makeFloors(Height height, int personCount, RungGenerator rungGenerator) {
-        return IntStream.range(0, height.getHeight())
+    private int parseAndValidateHeight(String inputHeight) {
+        int height = changeInputToInteger(inputHeight);
+        validateHeightRange(height);
+        return height;
+    }
+
+    private int changeInputToInteger(String input) {
+        try {
+            return Integer.parseInt(input);
+        } catch (NumberFormatException e) {
+            throw new NumberFormatException("숫자로 입력을 변환할 수 없습니다.");
+        }
+    }
+
+    private void validateHeightRange(int number) {
+        if (number < MIN_HEIGHT_RANGE || MAX_HEIGHT_RANGE < number) {
+            throw new IllegalArgumentException(
+                    "입력된 높이는 " + MIN_HEIGHT_RANGE + " 이상,"
+                            + " " + MAX_HEIGHT_RANGE + " 이하여야 합니다. 입력값 : " + number);
+        }
+    }
+
+    private List<Floor> makeFloors(int height, int personCount, RungGenerator rungGenerator) {
+        return IntStream.range(0, height)
                 .mapToObj(currentHeight -> new Floor(rungGenerator, personCount))
                 .toList();
     }
